Preliminary fixes for issues #192 and #193. Changed print statements to be debugger statements. Corrected formatting.
For #192: We remove subjects from the subject list as we encounter them, rather than storing up values to discard later.

For #193: Pattern we're examining is an operator, we examine what operator it is. If it is the same operator as is being applied to sublists, we extract its arguments. If we find this arguments in the subject list, we discard them and remove them from the operator’s argument list.