Changed add_bind and get_mandatory_bindings in SequenceMatcher to punt on variables with guards.
Reverted logic used in ACMatcher to select between SequenceMatcher.get_mandatory_bindings and ACMatcher.get_mandatory_bindings

Trying out a way of handling ANY typed variables. If an atom is not yet bound, there is no issue; the match and bind will occur as usual. If a binding already exists for a variable and an ANY is encountered for the current binding or the thing to be bound to, don't change anything. Just return the bindings as they are.